[KARAF-6362][karaf-maven-plugin] client goal:only one command in script file could be executed
(cherry picked from commit 2f966c675cd478f4cd0a5b6b4cf938a9f0064b88)

+++ tooling/karaf-maven-plugin/src/main/java/org/apache/karaf/tooling/client/CommandDescriptor.java
@@ -18,14 +18,14 @@ package org.apache.karaf.tooling.client;
 
 public class CommandDescriptor {
 
-    private int rank;
+    private double rank;
     private String command;
 
-    public int getRank() {
+    public double getRank() {
         return rank;
     }
 
-    public void setRank(int rank) {
+    public void setRank(double rank) {
         this.rank = rank;
     }
